# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PeoplePay360 API
  description: 'HR & Payroll Backend — Phase 7: Payruns & Payslips'
  version: 0.0.10
paths:
...
  /api/v1/employees/me/time-off/allocations:
    get:
      tags:
      - Employees
      summary: Get current logged-in employee time off allocations
      description: List all leave allocations for the currently authenticated employee.
      operationId: get_my_time_off_allocations_api_v1_employees_me_time_off_allocations_get
      security:
      - HTTPBearer: []
      parameters:
      - name: time_off_type_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Filter by leave type ID
          title: Time Off Type Id
        description: Filter by leave type ID
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/AllocationStatus'
          - type: 'null'
          description: Filter by allocation status
          title: Status
        description: Filter by allocation status
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeOffAllocationResponse'
                title: Response Get My Time Off Allocations Api V1 Employees Me Time
                  Off Allocations Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/me/time-off/requests:
    get:
      tags:
      - Employees
      summary: Get current logged-in employee time off requests
      description: List all leave requests for the currently authenticated employee.
      operationId: get_my_time_off_requests_api_v1_employees_me_time_off_requests_get
      security:
      - HTTPBearer: []
      parameters:
      - name: time_off_type_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Filter by leave type ID
          title: Time Off Type Id
        description: Filter by leave type ID
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/TimeOffRequestStatus'
          - type: 'null'
          description: Filter by request status
          title: Status
        description: Filter by request status
      - name: from_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          description: Filter requests ending on or after this date
          title: From Date
        description: Filter requests ending on or after this date
      - name: to_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          description: Filter requests starting on or before this date
          title: To Date
        description: Filter requests starting on or before this date
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeOffRequestResponse'
                title: Response Get My Time Off Requests Api V1 Employees Me Time
                  Off Requests Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Employees
      summary: Submit leave request for current logged-in employee
      description: Submit a new leave request for the currently authenticated employee.
      operationId: create_my_time_off_request_api_v1_employees_me_time_off_requests_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeOffRequestCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeOffRequestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/me/time-off/balances:
    get:
      tags:
      - Employees
      summary: Get current logged-in employee time off balances
      description: Calculate and return aggregated leave balances for the currently
        authenticated employee.
      operationId: get_my_time_off_balances_api_v1_employees_me_time_off_balances_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeOffBalanceResponse'
      security:
      - HTTPBearer: []
...
  /api/v1/employees/{employee_id}/time-off/allocations:
    get:
      tags:
      - Employees
      summary: Get all time off allocations for an employee
      description: 'List all leave allocations for a specific employee.

        Allowed for HR/Admin, or the employee themselves.'
      operationId: get_employee_time_off_allocations_api_v1_employees__employee_id__time_off_allocations_get
      security:
      - HTTPBearer: []
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: integer
          title: Employee Id
      - name: time_off_type_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Filter by leave type ID
          title: Time Off Type Id
        description: Filter by leave type ID
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/AllocationStatus'
          - type: 'null'
          description: Filter by allocation status
          title: Status
        description: Filter by allocation status
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeOffAllocationResponse'
                title: Response Get Employee Time Off Allocations Api V1 Employees  Employee
                  Id  Time Off Allocations Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/{employee_id}/time-off/requests:
    get:
      tags:
      - Employees
      summary: Get all time off requests for an employee
      description: 'List all leave requests for a specific employee.

        Allowed for HR/Admin, or the employee themselves.'
      operationId: get_employee_time_off_requests_api_v1_employees__employee_id__time_off_requests_get
      security:
      - HTTPBearer: []
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: integer
          title: Employee Id
      - name: time_off_type_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Filter by leave type ID
          title: Time Off Type Id
        description: Filter by leave type ID
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/TimeOffRequestStatus'
          - type: 'null'
          description: Filter by request status
          title: Status
        description: Filter by request status
      - name: from_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          description: Filter requests ending on or after this date
          title: From Date
        description: Filter requests ending on or after this date
      - name: to_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          description: Filter requests starting on or before this date
          title: To Date
        description: Filter requests starting on or before this date
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeOffRequestResponse'
                title: Response Get Employee Time Off Requests Api V1 Employees  Employee
                  Id  Time Off Requests Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/employees/{employee_id}/time-off/balances:
    get:
      tags:
      - Employees
      summary: Get employee leave balances
      description: 'Calculate and return aggregated leave balances for a specific
        employee.

        Allowed for HR/Admin, or the employee themselves.'
      operationId: get_employee_time_off_balances_api_v1_employees__employee_id__time_off_balances_get
      security:
      - HTTPBearer: []
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: integer
          title: Employee Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeOffBalanceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/time-off/types:
    get:
      tags:
      - Time Off
      summary: List all time off types
      operationId: list_types_api_v1_time_off_types_get
      security:
      - HTTPBearer: []
      parameters:
      - name: include_inactive
        in: query
        required: false
        schema:
          type: boolean
          description: Include deactivated leave types
          default: false
          title: Include Inactive
        description: Include deactivated leave types
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeOffTypeResponse'
                title: Response List Types Api V1 Time Off Types Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Time Off
      summary: Create a new time off type (HR/Admin)
      operationId: create_type_api_v1_time_off_types_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeOffTypeCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeOffTypeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/time-off/types/{id}:
    get:
      tags:
      - Time Off
      summary: Get time off type details
      operationId: get_type_api_v1_time_off_types__id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeOffTypeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Time Off
      summary: Update time off type (HR/Admin)
      operationId: update_type_api_v1_time_off_types__id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeOffTypeUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeOffTypeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Time Off
      summary: Soft-deactivate time off type (HR/Admin)
      operationId: delete_type_api_v1_time_off_types__id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeOffTypeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/time-off/requests/{id}/approve:
    post:
      tags:
      - Time Off
      summary: Approve leave request (HR/Admin only; row-level locked balance deduction)
      operationId: approve_request_api_v1_time_off_requests__id__approve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeOffRequestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AllocationStatus:
      type: string
      enum:
      - DRAFT
      - APPROVED
      - ACTIVE
      - EXPIRED
      - CANCELLED
      title: AllocationStatus
      description: 'Lifecycle status of a leave allocation grant.


        States:

        - DRAFT: Draft allocation under review.

        - APPROVED: Approved allocation ready for use.

        - ACTIVE: Currently active and eligible for consumption.

        - EXPIRED: Past the valid_to boundary.

        - CANCELLED: Revoked/cancelled allocation.'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TimeOffBalanceResponse:
      properties:
        employee_id:
          anyOf:
          - type: string
            format: uuid
          - type: integer
          - type: string
          title: Employee Id
        balances:
          items:
            $ref: '#/components/schemas/TimeOffBalanceItem'
          type: array
          title: Balances
      type: object
      required:
      - employee_id
      - balances
      title: TimeOffBalanceResponse
...
    TimeOffRequestStatus:
      type: string
      enum:
      - PENDING
      - APPROVED
      - REFUSED
      - CANCELLED
      title: TimeOffRequestStatus
      description: 'Lifecycle status of an employee leave request.


        States:

        - PENDING: Submitted and awaiting manager/HR review.

        - APPROVED: Leave approved and balance deducted.

        - REFUSED: Rejected by reviewer with mandatory refusal reason.

        - CANCELLED: Cancelled by employee or HR (restoring balance if previously
        approved).'
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer